+++ examples/custom_checkers/e9980_invalid_precondition_syntax.py
@@ -2,11 +2,10 @@
 
 
 def demo_function(x: int, y: int) -> int:
-    """
-    Demonstrates e9980_invalid_precondition_syntax
+    """Demonstrates e9980_invalid_precondition_syntax
 
     Preconditions:
        - x !== 0  # error on this line
        - y != 0  # no error on this line
     """
-    return x//y
+    return x // y